refactor: extract validateDefaultArgs to args.mts to avoid duplication
default-poll.mts had an inline arg-validation loop structurally identical
to the base branch's default-iterate.mts; SonarCloud flagged it as
duplicated new code. The loop is now a shared helper in args.mts that
both (and any future) default-dispatch validators can reuse.
